Split out linting
- Move into tools directory
- Put linting into new go package mdlint
- Remove duplication between implementation, tests and client usage
- Use relative import paths
