+++ Makefile
@@ -8,7 +8,7 @@ aspell:
 	find service-manual -type f -name '*.md' -exec aspell -d, --master=british --mode=sgml -c '{}' \;
 
 test:
-	go test lint_test.go
+	go test tools/lint/mdlint/mdlint_test.go
 
 help:
 	@echo "Please use 'make <target>' where <target> is one of"
